refactor: merge `useStore` hook and `createStoreContext`
* in order to inject store to components and
  use the store in the `Calendar` instance same time,
  the store should be injected through the Provider component.
* `useStore` hook doesn't have to be separated.
