+++ apps/calendar/src/store/context.ts
@@ -1,38 +1,115 @@
 import { createContext, createElement, FunctionComponent } from 'preact';
-import { useContext, useMemo, useRef } from 'preact/hooks';
+import { useContext, useEffect, useLayoutEffect, useMemo, useReducer, useRef } from 'preact/hooks';
 
-import { EqualityChecker, StateSelector, StateWithActions, StoreHooks, UseStore } from '@t/store';
+import { isNil, isUndefined } from '@src/util/utils';
+
+import { EqualityChecker, InternalStoreAPI, StateSelector, StateWithActions } from '@t/store';
+
+/**
+ * Inspired by Zustand
+ *
+ * See more: https://github.com/pmndrs/zustand
+ */
+
+const isSSR = isUndefined(window) || !window.navigator;
+
+const useIsomorphicLayoutEffect = isSSR ? useEffect : useLayoutEffect;
 
 export function createStoreContext<State extends StateWithActions>() {
-  type StoreContextType = StoreHooks<State>;
-  const StoreContext = createContext<StoreContextType | null>(null);
+  const StoreContext = createContext<InternalStoreAPI<State> | null>(null);
 
-  const Provider: FunctionComponent<{
-    storeCreator: () => StoreContextType;
-  }> = ({ storeCreator, children }) => {
-    const storeRef = useRef<StoreContextType>();
-
-    if (!storeRef.current) {
-      storeRef.current = storeCreator();
-    }
-
-    return createElement(StoreContext.Provider, { value: storeRef.current, children });
+  const StoreProvider: FunctionComponent<{ store: InternalStoreAPI<State> }> = ({
+    children,
+    store,
+  }) => {
+    return createElement(StoreContext.Provider, { value: store, children });
   };
 
-  const useStore: UseStore<State> = <StateSlice>(
-    selector?: StateSelector<State, StateSlice>,
-    equalityFn: EqualityChecker<State> = Object.is
+  const useStore = <StateSlice>(
+    selector: StateSelector<State, StateSlice>,
+    equalityFn: EqualityChecker<StateSlice> = Object.is
   ) => {
     const storeCtx = useContext(StoreContext);
 
-    if (!storeCtx?.useStore) {
+    if (isNil(storeCtx)) {
       throw new Error('StoreProvider is not found');
     }
 
-    // we cannot call `storeCtx.useStore` in condition because it violates the rules of hook.
-    // `selector` type should be `StateSelector<State, StateSlice>`
-    // but TS cannot infer the type of `selector` clearly.
-    return storeCtx.useStore(selector as StateSelector<State, State>, equalityFn);
+    // a little trick to invoke re-render to notify hook consumers(usually components)
+    const [, notify] = useReducer((notifyCount) => notifyCount + 1, 0) as [never, () => void];
+
+    const state = storeCtx.getState();
+    const stateRef = useRef(state);
+    const selectorRef = useRef(selector);
+    const equalityFnRef = useRef(equalityFn);
+    const hasErrorRef = useRef(false);
+    // `null` can be a valid state slice.
+    const currentSliceRef = useRef<StateSlice | undefined>();
+
+    if (isUndefined(currentSliceRef.current)) {
+      currentSliceRef.current = selector(state);
+    }
+
+    let newStateSlice: StateSlice | undefined;
+    let hasNewStateSlice = false;
+
+    const shouldGetNewSlice =
+      stateRef.current !== state ||
+      selectorRef.current !== selector ||
+      equalityFnRef.current !== equalityFn ||
+      hasErrorRef.current;
+    if (shouldGetNewSlice) {
+      newStateSlice = selector(state);
+      hasNewStateSlice = !equalityFn(currentSliceRef.current, newStateSlice);
+    }
+
+    useIsomorphicLayoutEffect(() => {
+      if (hasNewStateSlice) {
+        currentSliceRef.current = newStateSlice as StateSlice;
+      }
+
+      stateRef.current = state;
+      selectorRef.current = selector;
+      equalityFnRef.current = equalityFn;
+      hasErrorRef.current = false;
+    });
+
+    // NOTE: There is edge case that state is changed before subscription
+    const stateBeforeSubscriptionRef = useRef(state);
+    useIsomorphicLayoutEffect(() => {
+      const listener = () => {
+        try {
+          const nextState = storeCtx.getState();
+          const nextStateSlice = selectorRef.current(nextState);
+
+          const shouldUpdateState = !equalityFnRef.current(
+            currentSliceRef.current as StateSlice,
+            nextStateSlice
+          );
+          if (shouldUpdateState) {
+            stateRef.current = nextState;
+            currentSliceRef.current = newStateSlice;
+
+            notify();
+          }
+        } catch (e) {
+          // This will be rarely happened, unless we don't pass the arguments to actions properly.
+          // eslint-disable-next-line no-console
+          console.error('[toastui-calendar] failed to update state', e.message);
+          hasErrorRef.current = true;
+          notify();
+        }
+      };
+
+      const unsubscribe = storeCtx.subscribe(listener);
+      if (storeCtx.getState() !== stateBeforeSubscriptionRef.current) {
+        listener();
+      }
+
+      return unsubscribe;
+    }, []);
+
+    return hasNewStateSlice ? (newStateSlice as StateSlice) : currentSliceRef.current;
   };
 
   /**
@@ -42,15 +119,15 @@ export function createStoreContext<State extends StateWithActions>() {
   const useInternalStore = () => {
     const storeCtx = useContext(StoreContext);
 
-    if (!storeCtx) {
+    if (isNil(storeCtx)) {
       throw new Error('StoreProvider is not found');
     }
 
-    return useMemo(() => storeCtx.internalStore, [storeCtx]);
+    return useMemo(() => storeCtx, [storeCtx]);
   };
 
   return {
-    Provider,
+    StoreProvider,
     useStore,
     useInternalStore,
   };
